Playing with Kodein and sqlDelight

+++ shared/src/commonMain/kotlin/com/merseyside/testsharedproject/data/db/coinPair/CoinPairEntity.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CoinPairEntity(
     val name: String,
-    val high: Float,
-    val low: Float,
-    val avg: Float
+    val high: Double,
+    val low: Double,
+    val avg: Double
 )
